refactor(policies): collapse propose/update/remove into one manage-policies script
Replaces the three single-purpose built-in policies (propose-policy,
update-policy, remove-policy) with a single manage-policies built-in
that exposes `add`, `update`, and `remove` as subcommands. Read access
stays at `clawmini-lite requests show <name>` — the policy-script flow
always requires approval, so wiring reads through it would be wrong.

Each subcommand keeps the same behavior as the previous standalone
script (refuse-overwrite on add, refuse-missing on update/remove,
--disable-builtin to opt out of a built-in). Ports the e2e suites and
the workspace.test.ts resolver tests onto the new name; updates the
clawmini-requests and run-host skill docs so agents reach for the new
script.
