+++ templates/skills/run-host/SKILL.md
@@ -1,18 +1,18 @@
 ---
 name: run-host
-description: Last-resort escape hatch for running arbitrary shell commands on the host. Prefer proposing a dedicated policy first.
+description: Last-resort escape hatch for running arbitrary shell commands on the host. Prefer registering a dedicated policy first.
 ---
 
 # Run Host
 
-**Before using this, propose a dedicated policy.** `run-host` is a generic escape hatch that forwards a command string to `sh -c`. It exists so you are never fully blocked, but it is the wrong tool for anything you will do more than once.
+**Before using this, register a dedicated policy.** `run-host` is a generic escape hatch that forwards a command string to `sh -c`. It exists so you are never fully blocked, but it is the wrong tool for anything you will do more than once.
 
-For any task you can describe — running tests, installing a package, calling a specific API — use `propose-policy` to register a named policy instead. Dedicated policies are easier for the user to review, can be auto-approved when safe, and give you a clear skill to call next time. Reach for `run-host` only when:
+For any task you can describe — running tests, installing a package, calling a specific API — use `manage-policies add` to register a named policy instead. Dedicated policies are easier for the user to review, can be auto-approved when safe, and give you a clear skill to call next time. Reach for `run-host` only when:
 
 - You need a genuinely one-off exploratory command (e.g. inspecting an unfamiliar file), **or**
 - You truly cannot express the task as a reusable policy.
 
-Because it can do anything the host shell can do, the user will be asked to approve every invocation, and repeated `run-host` requests for similar tasks are a signal that you should have proposed a policy.
+Because it can do anything the host shell can do, the user will be asked to approve every invocation, and repeated `run-host` requests for similar tasks are a signal that you should have registered a policy.
 
 ## Usage
