+++ tsdown.config.ts
@@ -17,25 +17,7 @@ export default defineConfig([
     inlineOnly: false,
   },
   {
-    entry: ['src/cli/propose-policy.ts'],
-    format: 'esm',
-    dts: true,
-    clean: false,
-    outDir: 'dist/cli',
-    noExternal: [/(.*)/],
-    inlineOnly: false,
-  },
-  {
-    entry: ['src/cli/update-policy.ts'],
-    format: 'esm',
-    dts: true,
-    clean: false,
-    outDir: 'dist/cli',
-    noExternal: [/(.*)/],
-    inlineOnly: false,
-  },
-  {
-    entry: ['src/cli/remove-policy.ts'],
+    entry: ['src/cli/manage-policies.ts'],
     format: 'esm',
     dts: true,
     clean: false,
